refactor(workers): isolate scheduler state service
