+++ apps/server/workers/src/crons/youtube/cron.youtube-status.service.spec.ts
@@ -8,8 +8,8 @@ import { PublishEventWebhookService } from '@api/services/webhook-client/webhook
 import { PostStatus } from '@genfeedai/enums';
 import { LoggerService } from '@libs/logger/logger.service';
 import { Test, TestingModule } from '@nestjs/testing';
-import { SchedulerPublishStateService } from '@workers/crons/posts/scheduler-publish-state.service';
 import { CronYoutubeStatusService } from '@workers/crons/youtube/cron.youtube-status.service';
+import { SchedulerPublishStateService } from '@workers/services/scheduler-publish-state.service';
 
 describe('CronYoutubeStatusService', () => {
   let service: CronYoutubeStatusService;
