+++ apps/server/workers/src/crons/youtube/cron.youtube.module.ts
@@ -5,11 +5,11 @@ import { YoutubeModule } from '@api/services/integrations/youtube/youtube.module
 import { WebhookClientModule } from '@api/services/webhook-client/webhook-client.module';
 import { PrismaModule } from '@libs/prisma/prisma.module';
 import { forwardRef, Module } from '@nestjs/common';
-import { SchedulerPublishStateService } from '@workers/crons/posts/scheduler-publish-state.service';
 import { CronYoutubeAnalyticsService } from '@workers/crons/youtube/cron.youtube-analytics.service';
 import { CronYoutubeMessagesService } from '@workers/crons/youtube/cron.youtube-messages.service';
 import { CronYoutubeStatusService } from '@workers/crons/youtube/cron.youtube-status.service';
 import { WorkersQueuesModule } from '@workers/queues/queues.module';
+import { SchedulerPublishStateService } from '@workers/services/scheduler-publish-state.service';
 
 @Module({
   imports: [
